*** empty log message ***

+++ trunk/pymol/layer1/Scene.h
@@ -94,7 +94,7 @@ void ScenePrepareExit(PyMOLGlobals *G);
 void SceneGetViewNormal(PyMOLGlobals *G,float *v);
 void SceneClipSet(PyMOLGlobals *G,float front,float back);
 void SceneGetView(PyMOLGlobals *G,SceneViewType view);
-void SceneSetView(PyMOLGlobals *G,SceneViewType view,int quiet,int animate);
+void SceneSetView(PyMOLGlobals *G,SceneViewType view,int quiet,float animate);
 
 void SceneToViewElem(PyMOLGlobals *G,CViewElem *elem);
 void SceneFromViewElem(PyMOLGlobals *G,CViewElem *elem);
